Created modal dialog for newly found geodes

+++ iFind/IntroPageViewController.m
@@ -161,7 +161,7 @@
     }
     else if (i == 3){
         SignUpViewController * SignUpController = [sbptr instantiateViewControllerWithIdentifier:@"SignUp"];
-        SignUpController.pioneerRank = self.responseRank;
+        SignUpController.pioneerRank = (int)self.responseRank;
         SignUpController.responseParentNode = self.responseParentNode;
         return SignUpController;
     }
